api: add worker
The worker runs periodically and will perform some tasks. For now,
just a print statement to verify if its running.
