+++ .env
@@ -35,6 +35,10 @@ SHELLHUB_DOMAIN=localhost
 # NOTICE: Redis is required
 SHELLHUB_STORE_CACHE=false
 
+# Set worker's schedule
+# NOTICE: The format is the same as the Go implementation of https://pkg.go.dev/github.com/robfig/cron
+SHELLHUB_WORKER_SCHEDULE=@daily
+
 # Recording session host
 SHELLHUB_RECORD_URL=api:8080
